move ProductController extensions from project-base to the framework package
