+++ project-base/app/src/Controller/Admin/ProductController.php
@@ -24,43 +24,5 @@ use Symfony\Component\Routing\Annotation\Route;
  */
 class ProductController extends BaseProductController
 {
-    /**
-     * @param \Symfony\Component\HttpFoundation\Request $request
-     * @return \Symfony\Component\HttpFoundation\Response
-     */
-    #[Route(path: '/product/edit/catnum-exists')]
-    public function catnumExistsAction(Request $request): Response
-    {
-        $catnum = $request->get('catnum');
-        $currentProductCatnum = $request->get('currentProductCatnum');
 
-        if ($catnum === null || $catnum === $currentProductCatnum) {
-            return new JsonResponse(false);
-        }
-
-        $productByCatnum = $this->productFacade->findByCatnum($catnum);
-
-        return new JsonResponse($productByCatnum !== null);
-    }
-
-    /**
-     * This route is used by GrapesJS to load Names of products
-     *
-     * @param \Symfony\Component\HttpFoundation\Request $request
-     * @return \Symfony\Component\HttpFoundation\JsonResponse
-     */
-    #[Route(path: '/product/names-by-catnums', methods: ['post'], condition: 'request.isXmlHttpRequest()')]
-    public function productNamesByCatnumsAction(Request $request): JsonResponse
-    {
-        $catnums = $request->get('catnums');
-
-        $response = [];
-        $products = $this->productFacade->findAllByCatnums($catnums);
-
-        foreach ($products as $product) {
-            $response[$product->getCatnum()] = $product->getName();
-        }
-
-        return new JsonResponse($response);
-    }
 }
